Added decoders for _mpl, VidM and InPr fields
Decoding the InPr bytes is a bit hard since Atem Mini doesn't
have enough difference in the input properties to discern the fields.
For example the only inputs are all HDMI and thus the port type that's
selected is 0x02, but this lines up perfectly with the bitfield for
port options being HDMI only, which is also 0x02. I need a more
complex mixer to see which of the 2 bytes is the select and which
is the available ports.