Retrieve proper mix and ftb rate
The frame count for the transitions don't seem to be actually based on
the framerate the mixer is running at. A 25 frame transition takes one
second in 1080p25 mode, it also takes a second in 1080p50 mode. It seems
like the rates get rounded to either 25 or 30 fps and thats used to
calculate the real time for frame durations in the transitions.