Some sources are not available on both AUX1 and AUX2
The Atem Mini Extreme has got two AUX outputs ("HDMI OUT 1" and "HDMI
OUT 2"). Two of the inputs, HDMI1 and HDMI2, support a special "direct
mode" where they can be passed through to an HDMI output at a lower
latency (perhaps just via a HDMI retimer?). This is similar to the
original Atem Mini series where this was only supported for HDMI1. Now,
on the Extreme, HDMI1 can be low-latency-put to HDMI OUT 1, and the
HDMI2 can be passed through to HDMI OUT 2. This is a problem because the
code would not know the difference, and there was apparently just one
bit for the "can be sent to AUX" flag.

The catch is that on the Atem Mini Pro at least, the "can be sent to
HDMI OUT 1" bit is always 0, even on the newest FW. On the Atem Mini
Extreme, I have not seen any source with available_aux == 1 and
available_aux1 == available_aux2 == 0. The only sources where
available_aux1 != available_aux2 are the DIR1 and DIR2 for the direct
pass-through.

Signed-off-by: Jan Kundrát <jkt@jankundrat.com>