Drop retransmission packets of received packets
Sometimes the mixer will send data faster than python will ACK
it causing the hardware to send retransmissions with duplicate
data. This will keep track of the 128 last received packet ids
and drop the retransmission packets if we did actually retrieve
it correctly previously.

The number 128 was chosen because the hardware seems to allow
about 64 unacked frames in flight.