proxy: do not assume all frontends "bind" somewhere
I'm writing a frontend which actually *connects* to an external server,
so let's not make any assumptions here. The only downside is that we
cannot easily print out what section of the config file is getting
handled before we initialize its handler. I think this is better than
calling a parameter for MQTT server's hostname to connect to "bind".

Signed-off-by: Jan Kundrát <jkt@jankundrat.com>