proxy: mqtt: allow arbitrary writes
Since there are typically no server-side MQTT ACLs, this has to be
enabled explicitly.

There's some duplication of code between this implementation and the one
in the HTTP backend. The most straightforward way of fixing that is to
convert these early returns into proper exceptions, but I'm too lazy for
this and I'm gonna be using this feature tomorrow :).

Signed-off-by: Jan Kundrát <jkt@jankundrat.com>