proxy: mqtt: make subscribe topic configurable
The proxy now subscribes to the same topic as set for sending by default
if allow-writes is enabled. To have a seperate subcription as the
sending topic use the topic-subscribe setting.

The protocol is switched to MQTTv5 to be able to use NoLocal which makes
the broker not send the messages back to the sender when the sending and
receiving topic is the same.