openswitcher: handle exception caused by disconnecting on initial sync
The connection closing in the middle of the initial sync leaves a lot
of GTK evens queued that cause exceptions down the line. By checking
for the connection state here it will make the state update
exceptions non-critical when there's no connection anymore.

This allows the UI to function enough to create a new connection